new set of tests applied

+++ ft_strchr.c
@@ -6,15 +6,15 @@
 /*   By: acioalai <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/01 11:06:01 by acioalai          #+#    #+#             */
-/*   Updated: 2015/11/04 08:03:51 by acioalai         ###   ########.fr       */
+/*   Updated: 2015/11/07 02:47:55 by acioalai         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "header.h"
+#include "libft.h"
 
 char	*ft_strchr(const char *s, int c)
 {
-	int		i;
+	size_t	i;
 
 	i = 0;
 	while (i <= (ft_strlen(s) + 1))
