+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: acioalai <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/10/24 12:40:25 by acioalai          #+#    #+#             */
-/*   Updated: 2015/11/04 07:59:12 by acioalai         ###   ########.fr       */
+/*   Updated: 2015/11/07 02:47:13 by acioalai         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "header.h"
+#include "libft.h"
 
 size_t	ft_strlen(const char *str)
 {
-	int		i;
+	size_t		i;
 
 	i = 0;
 	while (str[i])
